feat: cross-platform process termination for test server
Split TerminateProcessGroup into platform-specific files:
- Unix: SIGTERM to process group, SIGKILL fallback
- Windows: taskkill /T /F /PID for process tree kill
